finished business layer

+++ data/src/D2CFL.Data.FantasyLeague/Contract/PlayerStatisticsEntity.cs
@@ -9,7 +9,7 @@ namespace D2CFL.Data.FantasyLeague.Contract
         public int TotalKills { get; set; }
         public int TotalAssists { get; set; }
         public int TotalDeaths { get; set; }
-        public int TotalPoints { get; set; }
+        public double TotalPoints { get; set; }
         public double AverageKills { get; set; }
         public double AverageAssists { get; set; }
         public double AverageDeaths { get; set; }
